rename: 로그인 타입 구분하는 enum class 경로 변경
