+++ app/src/main/java/com/tellingus/tellingme/presentation/ui/feature/login/Enum.kt
@@ -1,0 +1,10 @@
+package com.tellingus.tellingme.presentation.ui.feature.login
+
+enum class LoginType(name: String) {
+    KAKAO("kakao")
+}
+
+enum class IsAuto(name: String) {
+    MANUAL("manual"),
+    AUTO("auto")
+}
