+++ app/src/main/java/com/tellingus/tellingme/presentation/viewmodel/LoginViewModel.kt
@@ -76,12 +76,3 @@ class LoginViewModel @Inject constructor(
         }
     }
 }
-
-
-enum class LoginType(name: String) {
-    KAKAO("kakao")
-}
-enum class IsAuto(name: String) {
-    MANUAL("manual"),
-    AUTO("auto")
-}
